Move cluster topology calculations to a separate file.
