+++ redis/src/cluster_topology.rs
@@ -1,0 +1,68 @@
+use crate::{cluster::get_connection_addr, cluster_routing::Slot, RedisResult, TlsMode, Value};
+
+// Parse slot data from raw redis value.
+pub(crate) fn parse_slots(raw_slot_resp: Value, tls: Option<TlsMode>) -> RedisResult<Vec<Slot>> {
+    // Parse response.
+    let mut result = Vec::with_capacity(2);
+
+    if let Value::Array(items) = raw_slot_resp {
+        let mut iter = items.into_iter();
+        while let Some(Value::Array(item)) = iter.next() {
+            if item.len() < 3 {
+                continue;
+            }
+
+            let start = if let Value::Int(start) = item[0] {
+                start as u16
+            } else {
+                continue;
+            };
+
+            let end = if let Value::Int(end) = item[1] {
+                end as u16
+            } else {
+                continue;
+            };
+
+            let mut nodes: Vec<String> = item
+                .into_iter()
+                .skip(2)
+                .filter_map(|node| {
+                    if let Value::Array(node) = node {
+                        if node.len() < 2 {
+                            return None;
+                        }
+
+                        let ip = if let Value::BulkString(ref ip) = node[0] {
+                            String::from_utf8_lossy(ip)
+                        } else {
+                            return None;
+                        };
+                        if ip.is_empty() {
+                            return None;
+                        }
+
+                        let port = if let Value::Int(port) = node[1] {
+                            port as u16
+                        } else {
+                            return None;
+                        };
+                        // This is only "stringifying" IP addresses, so `TLS parameters` are not required
+                        Some(get_connection_addr(ip.into_owned(), port, tls, None).to_string())
+                    } else {
+                        None
+                    }
+                })
+                .collect();
+
+            if nodes.is_empty() {
+                continue;
+            }
+
+            let replicas = nodes.split_off(1);
+            result.push(Slot::new(start, end, nodes.pop().unwrap(), replicas));
+        }
+    }
+
+    Ok(result)
+}
